seperate out artifactmap functions into seperate file, change types quiet a bit with e2e still passing
