+++ src/clientSideScene/sceneEntities.ts
@@ -95,10 +95,8 @@ import { createGridHelper, orthoScale, perspScale } from './helpers'
 import { Models } from '@kittycad/lib'
 import { uuidv4 } from 'lib/utils'
 import { SegmentOverlayPayload, SketchDetails } from 'machines/modelingMachine'
-import {
-  ArtifactMapCommand,
-  EngineCommandManager,
-} from 'lang/std/engineConnection'
+import { EngineCommandManager } from 'lang/std/engineConnection'
+import { ArtifactMapCommand } from 'lang/std/artifactMap'
 import {
   getRectangleCallExpressions,
   updateRectangleSketch,
